Extract `BackgroundLayers` component
