+++ ember/app/components/base-map.hbs
@@ -1,3 +1,4 @@
 <div ...attributes {{set-map-target this.map}}>
+  <BackgroundLayers @map={{map}} />
   <LayerSwitcher @map={{map}} />
 </div>
