+++ ember/app/components/takeoffs-map.hbs
@@ -1,4 +1,5 @@
 <div class="relative" ...attributes {{set-map-target this.map}}>
+  <BackgroundLayers @map={{map}} />
   <LayerSwitcher @map={{map}} />
   <TakeoffsLayer @map={{map}} @locations={{locations}} />
   <MapClickHandler @map={{map}} />
